imap: move over imapclient.List{Options,Data}
